+++ components/MainNavigation/BottomTabNavigator./index.js
@@ -2,12 +2,12 @@ import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import * as React from 'react';
 
 import TabBarIcon from '../../TabBarIcon';
-import HomeScreen from '../../../domains/HomeScreen';
+// import HomeScreen from '../../../domains/HomeScreen';
 import DataCollection from '../../../domains/DataCollection';
 import DataAnalysis from '../../../domains/DataAnalysis';
 
 const BottomTab = createBottomTabNavigator();
-const INITIAL_ROUTE_NAME = 'Home';
+const INITIAL_ROUTE_NAME = 'Data_Collection';
 
 export default function BottomTabNavigator({ navigation, route }) {
   navigation.setOptions({
@@ -17,14 +17,14 @@ export default function BottomTabNavigator({ navigation, route }) {
 
   return (
     <BottomTab.Navigator initialRouteName={INITIAL_ROUTE_NAME}>
-      <BottomTab.Screen
+      {/* <BottomTab.Screen
         name="Home"
         component={HomeScreen}
         options={{
           title: 'Get Started',
           tabBarIcon: ({ focused }) => <TabBarIcon focused={focused} name="md-home" />,
         }}
-      />
+      /> */}
       <BottomTab.Screen
         name="Data_Collection"
         component={DataCollection}
@@ -50,13 +50,15 @@ function getHeaderTitle(route) {
   const routeName = route.state?.routes[route.state.index]?.name ?? INITIAL_ROUTE_NAME;
 
   switch (routeName) {
-    case 'Home':
-      return 'Home';
+    // case 'Home':
+    //   return 'Home';
     case 'Data_Collection':
       return 'Data Collection';
     case 'Data_Analysis':
       return 'Data Analysis';
+    // default:
+    //   return 'Home';
     default:
-      return 'Home';
+      return 'Data Collection';
   }
 }