+++ domains/HomeScreen/index.js
@@ -25,17 +25,20 @@ const HomeScreen = (props) => {
       setTasks(result);
     });
   };
+
   const logOut = () => {
     retrieveSignOutFunction().then(() => {
       deleteData('credentials');
       deleteData('pincode');
       deleteData('organization');
+      deleteData('currentUser');
       navigation.navigate('Sign In');
     });
   };
+
   return (
     <View style={layout.screenContainer}>
-      <Header />
+      <Header logOut={logOut} />
       <ScrollView>
         <View style={layout.screenRow}>
           <Title>My Tasks</Title>